(External) password change for Shibboleth users
As one of the original authors of the Shibboleth Moodle auth plugin I propose the following change, which modifies the function can_change_password()  and adds change_password_url() to allow Shibboleth users to change their password in case there is an (external) password change URL defined in Moodle. If no such URL is defined, the behaviour is the same as without the proposed change.
This feature is primarily useful in a scenario where Moodle is used by users of one single organisation (or very few organisations) only.
Code is in use and was contributed by FH-HWZ.ch.